App.js: replace copy-paste code with loop
